// src/models/googleLocationDetails.ts
export interface googleLocationsDetails{
     formatted_address : string;
    // formatted_phone_number : string;
    // lat :number;
    // lng : number;
    // international_phone_number : string;
    // name : string;
    // website : string;
    // url : string;
    Street: string;
    State: string;
    City: string;
    Zip: string;
}